requirements.txt
flask
boto3
awscli
pypdf
langchain
langchain_aws
langchain_community
langchain-text-splitters
pyarrow
streamlit
faiss-cpu
sentence-transformers
python-dotenv
langchain-pinecone
transformers
tokenizers

-e .
